Squash commits of Slim Battery UI Customization
fb: add Battery Styles the new way 1/2

google draws now the battery icon which fits perfectly into our
concept. Beside that the batteryview is as well used on QS.
So we readd our battery icons (icon, percentage, icon with percentage
circle, circle perc, dot circle, dot circle perc) back and all are
drawn. As well the icon does not only change on statusbar as well
on QS it is reflected.

As usual color options are present for the User to style the icon

Change-Id: Iae9021af56a21b8d39efc5162186eb17ecc933e2

fb: show percentage on icon in bold and use for

only percent more width in statusbar and QS
As well show on percent with icon on fully charge not the bolt
and instead the value 100

Change-Id: I4e0b837ff6e217e1c9f78d967aa57c3e2f1bf2ad

fb: fix rounding issues and take density into account

we are a multiDpi rom...and on 4.2 and 4.3 already the text
in the circle battery icon was on lower dpi a small bit up. The original
author compensated this with just a +1 to the Y value. Thats not the right
way...calculate the current dpi into the offset and all is well.

We want a perfect 4.4...so this nitpicks even if we are talking about
0.5dp maybe should be fixed

Change-Id: I835aa20b82aea980eb79a4772adccf30b2d6f79f

TODO: fix quicksettings battery ui update as it didn't update itself

Change-Id: Iabd839987918445872f4c64317f290e1637694e3

Conflicts:
	packages/SystemUI/res/layout/status_bar.xml
	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBar.java
	packages/SystemUI/src/com/android/systemui/statusbar/phone/QuickSettings.java